+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-1.0.0 / 2017-02-06
-==================
+### 1.1.0 / 2017-05-01
+* Fix final module merge to improve efficiency and recalculate eigengenes after each merge 
+* Change program output to generate iteration-specific output files in easily accessible directory structure
+* add post-processing scripts
 
+### 1.0.0 / 2017-02-06
 * First numbered release
 * Version names will follow "Semantic Versioning" of X.Y.Z where:
     * Increment Z when you fix something